bie avancer sur la lib memoire

+++ src/solver/short_algo/algo.c
@@ -12,18 +12,19 @@
 
 #include "../../includes/all_includes.h"
 
-void destroy_finder(t_finder finder)
+void * destroy_finder(t_finder finder)
 {
 	free(finder->taken_room);
 	destroy_dll_func(&finder->working_path, dll_l_notfree_content);
 	destroy_dll_func(&finder->new_path, dll_l_notfree_content);
 	free(finder);
+	return (NULL);
 }
 
 t_move algo(t_cache cache, t_data data, t_map map)
 {
 	t_finder finder;
-	//	t_map path_map;
+	t_map path_map;
 	//	t_best_path best;
 	t_dll_l path_l;
 	(void) path_l;
@@ -37,23 +38,22 @@ t_move algo(t_cache cache, t_data data, t_map map)
 	/*------------------------------------*\
 		cherche les path
 	\*------------------------------------*/
-//	if (DEBUG->little == TRUE ||
-//		split_all_path(finder, map) == FALSE
-//	 )
+	if (DEBUG->little == TRUE ||
+		split_all_path(finder, map) == FALSE
+	 )
 	{
 		split_all_path(finder, map);
 		destroy_finder(finder);
 		finder = shorty_baby(cache, data, map);
 	}
-	//
-	//	if (finder->valid_path->length == 0)
-	//		return (NULL);
+	if (finder->valid_path->length == 0)
+		return (destroy_finder(finder));
 	//	// si return false --> faire :
 	//
 	//	/*------------------------------------*\
 //	    genere la map des path
 	//	\*------------------------------------*/
-	//	path_map = generate_path_map(data->room, finder->valid_path);
+		path_map = generate_path_map(data->room, finder->valid_path);
 	//
 	//	/*------------------------------------*\
 //	    trie les path
@@ -62,7 +62,7 @@ t_move algo(t_cache cache, t_data data, t_map map)
 	//	find_best_path(path_map, best);
 	//	return (new_move(data, &best->data, finder));
 	// clear finder
+	destroy_map(path_map);
 	destroy_finder(finder);
-
 	return (NULL);
 }
